+++ src/index.ts
@@ -5,6 +5,7 @@ import { cloudfunctionsv2Policies } from './cloudfunctionsv2'
 import { cloudrunPolicies } from './cloudrun'
 import { cloudrunv2Policies } from './cloudrunv2'
 import { computePolicies } from './compute'
+import { sqlPolicies } from './sql'
 import { storagePolicies } from './storage'
 
 const allPolicies = [
@@ -14,6 +15,7 @@ const allPolicies = [
     ...cloudrunPolicies,
     ...cloudrunv2Policies,
     ...computePolicies,
+    ...sqlPolicies,
     ...storagePolicies
 ]